Chapter 16 forms

+++ src/OC/PlatformBundle/Controller/AdvertController.php
@@ -11,7 +11,8 @@ use OC\PlatformBundle\Entity\Advert;
 use OC\PlatformBundle\Entity\Image;
 use OC\PlatformBundle\Entity\Application;
 use OC\PlatformBundle\Entity\AdvertSkill;
-
+use OC\PlatformBundle\Form\AdvertType;
+use OC\PlatformBundle\Form\AdvertEditType;
 
 class AdvertController extends Controller
 {
@@ -84,15 +85,27 @@ class AdvertController extends Controller
   //Page platform/add
   public function addAction(Request $request)
   {
-    $em = $this->getDoctrine()->getManager();
+    $advert = new Advert();
+    $form = $this->get('form.factory')->create(AdvertType::class, $advert);
 
     if($request->isMethod('POST'))
     {
-      $request->getSession()->getFlashBag()->add('notice', 'Annonce bien enregistrée');
-      return $this->redirectToRoute('oc_platform_view', array('id' => $advert->getId()));
+      $form->handleRequest($request);
+
+      if($form->isValid())
+      {
+        $em = $this->getDoctrine()->getManager();
+        $em->persist($advert);
+        $em->flush();
+
+        $request->getSession()->getFlashBag()->add('notice', 'Annonce bien enregistrée');
+        return $this->redirectToRoute('oc_platform_view', array('id' => $advert->getId()));
+      }
     }
 
-    return $this->render('@OCPlatform/Advert/add.html.twig');
+    return $this->render('@OCPlatform/Advert/add.html.twig',
+      array('form' => $form->createView())
+    );
   }
 
 
@@ -108,17 +121,29 @@ class AdvertController extends Controller
       throw new NotFoundHttpException("L'annonce avec l'id ".$id." n'existe pas.");
     }
 
+    $form = $this->get('form.factory')->create(AdvertEditType::class, $advert);
+
     if($request->isMethod('POST'))
     {
-      $request->getSession()->getFlashBag()->add('notice', 'Annonce bien modifiée');
-      return $this->redirectToRoute('oc_platform_view', array('id' => $advert->getId()));
+      $form->handleRequest($request);
+
+      if($form->isValid())
+      {
+        $em = $this->getDoctrine()->getManager();
+        $em->persist($advert);
+        $em->flush();
+
+        $request->getSession()->getFlashBag()->add('notice', 'Annonce bien modifiée');
+        return $this->redirectToRoute('oc_platform_view', array('id' => $advert->getId()));
+      }
     }
-    return $this->render('@OCPlatform/Advert/edit.html.twig', array('advert' => $advert));
+
+    return $this->render('@OCPlatform/Advert/edit.html.twig', array('form' => $form->createView()));
   }
 
 
   //Page platform/delete/id
-  public function deleteAction($id)
+  public function deleteAction(Request $request, $id)
   {
     $em = $this->getDoctrine()->getManager();
 
@@ -129,10 +154,19 @@ class AdvertController extends Controller
       throw new NotFoundHttpException("L'annonce avec l'id ".$id." n'existe pas.");
     }
 
-    $em->remove($advert);
-    $em->flush();
+    $form = $this->get('form.factory')->create();
 
-    return $this->render('@OCPlatform/Advert/delete.html.twig');
+    if($request->isMethod('POST') && $form->handleRequest($request))
+    {
+      $em->remove($advert);
+      $em->flush();
+      $request->getSession()->getFlashBag()->add('notice', 'Annonce bien supprimée');
+      return $this->redirectToRoute('oc_platform_homepage');
+    }
+
+    return $this->render('@OCPlatform/Advert/delete.html.twig',
+      array('advert' => $advert, 'form' => $form->createView())
+    );
   }
